Refactored custom theme
