+++ apps/web/src/theme/colors.ts
@@ -1,0 +1,10 @@
+const colors = {
+    brand: {
+        100: "#DFE0D7", // Background
+        300: "#E8E9E3", // CS Cream
+        500: "#00E046", // CS Chaman
+        900: "#003C00" // CS Wisdom
+    }
+}
+
+export default colors;
