+++ apps/web/src/theme/fonts.ts
@@ -1,0 +1,8 @@
+import "@fontsource/roboto";
+
+const fonts = {
+    heading: "VictorSerifTrial",
+    body: "Roboto"
+}
+
+export default fonts;
